improve clean-prefix

+++ scripts/clean-prefix.sh
@@ -13,17 +13,21 @@ source qcore/set-prefix.sh
     rmdir "$prefix" || true
 
     if [ -e "$prefix" ] ; then
-        echo "$prefix already exist, continue to build will erase all its contents."
         echo "Use ./scripts/qlat.sh to build Qlat only."
-        echo "Ctrl-C to stop."
-        for i in {10..0} ; do
-            echo -n "$i "
-            sleep 1;
+        echo "$prefix already exist, continue to build will erase all its contents."
+        while : ; do
+            echo "REALLY Want to continue? Crtl-C to interupt, type 'yes' to continue"
+            prompt=""
+            read prompt
+            if [ "$prompt" == "yes" ] ; then
+                break
+            fi
         done
-        echo
-        prefix_tmp=$(mktemp -d $prefix.tmp.XXXXX)
-        mv "$prefix" "$prefix_tmp"
-        rm -rf "$prefix_tmp" || true
+        if [ "$prompt" == "yes" ] ; then
+            prefix_tmp=$(mktemp -d $prefix.tmp.XXXXX)
+            mv "$prefix" "$prefix_tmp"
+            rm -rf "$prefix_tmp" || true
+        fi
     fi
 
     mkdir -p "$prefix"
